+++ wumgr/Wpf/WpfListSelectionPolicy.cs
@@ -1,0 +1,10 @@
+namespace wumgr.Wpf
+{
+    internal static class WpfListSelectionPolicy
+    {
+        public static bool CanSelectRows(WpfUpdateListKind list)
+        {
+            return list != WpfUpdateListKind.History;
+        }
+    }
+}